[DEV-174] Remove CryptoBuy + Refactoring
