+++ src/payment/models/bank-tx/bank-tx.entity.ts
@@ -2,7 +2,6 @@ import { IEntity } from 'src/shared/models/entity';
 import { Entity, Column, ManyToOne, OneToOne } from 'typeorm';
 import { BuyCrypto } from '../buy-crypto/entities/buy-crypto.entity';
 import { BuyFiat } from '../buy-fiat/buy-fiat.entity';
-import { CryptoBuy } from '../crypto-buy/crypto-buy.entity';
 import { BankTxBatch } from './bank-tx-batch.entity';
 
 export enum BankTxType {
@@ -142,9 +141,6 @@ export class BankTx extends IEntity {
 
   @ManyToOne(() => BankTxBatch, (batch) => batch.transactions, { nullable: true })
   batch: BankTxBatch;
-
-  @OneToOne(() => CryptoBuy, (buy) => buy.bankTx, { nullable: true })
-  cryptoBuy?: CryptoBuy;
 
   @OneToOne(() => BuyCrypto, (buyCrypto) => buyCrypto.bankTx, { nullable: true })
   buyCrypto?: BuyCrypto;
