--- src/payment/models/crypto-buy/crypto-buy.repository.ts
@@ -1,5 +1,0 @@
-import { EntityRepository, Repository } from 'typeorm';
-import { CryptoBuy } from './crypto-buy.entity';
-
-@EntityRepository(CryptoBuy)
-export class CryptoBuyRepository extends Repository<CryptoBuy> {}
