--- src/payment/models/crypto-buy/dto/create-crypto-buy.dto.ts
@@ -1,10 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsInt, IsNotEmpty } from 'class-validator';
-import { CryptoBuyDto } from './crypto-buy.dto';
-
-export class CreateCryptoBuyDto extends CryptoBuyDto {
-  @ApiProperty()
-  @IsNotEmpty()
-  @IsInt()
-  bankTxId: number;
-}
