--- src/payment/models/crypto-buy/enums/aml-check.enum.ts
@@ -1,5 +1,0 @@
-export enum AmlCheck {
-  PASS = 'Pass',
-  FAIL = 'Fail',
-  PENDING = 'Pending',
-}
